Employee's Face-Recognition Using-CNNs
This is a Face Recognition model using a CNN architecture with 5 layers of Conv2D, MaxPooling, BatchNormalization, and fully connected layers with ReLU activation to aggregate features. The optimizer used is Adam. After training with a batch size of 32, 20 epochs, and setting a threshold, the accuracy achieved is >0.90. Additionally, the system includes other features such as:

1. Input new employee image for training: Manual data augmentation with over 40 different types, which will automatically be generated and saved into the company's main folder using the input image name.

2. Delete Employee Image Folder

3. Delete Employee Image Folder (delete only images)

4. Create a new folder for the company

Note:

- For option 1, you must upload each image individually, because the system requires adding a new employee image this way (or you can change it to a folder path).

- Please review and adjust the directory paths.

- For this CNN model, you should have a minimum of 1500-2000 images, 50 classes, with around 40-50 images per class to avoid overfitting.

- It is recommended to use 3x4 or 4x6 images, as I will provide a source code for data preprocessing using YOLOv8 within this project.

This is a student project, so there might be some shortcomings. Please feel free to review and provide feedback. Thank you!