Employees Image Preprocessing using YoloV8
This is the image preprocessing step for employee photos using YOLOv8 for detection. You can use it directly for the Employee Face Recognition task or adapt it for other applications. The workflow is summarized as follows:

You can input either a single image file or a folder of images:

+ Single image file:
If the image is of an employee and meets the 3x4 or 4x6 aspect ratio requirements, it will be cropped using a recalculated bounding box that fits the face area only.
The cropped image will then be saved into a corresponding folder — either 3x4 or 4x6 — depending on the detected aspect ratio.
For images that do not contain a person or do not meet the expected dimensions, the model will either skip them or, if cropped, save them into a folder corresponding to their category (currently, only a demo for the fruit category in YOLOv8 is implemented — other categories will be added later).

+ Folder of images:
The behavior is the same. If the folder contains images that are not of people or do not meet the expected dimensions, they will be skipped or moved to the fruit folder.